+++ tests/test_split_empty_string.c
@@ -9,8 +9,7 @@ void test_impl()
   assert_string_equal(strings.strings[0], "");
   assert_signed_num_equal(strings.count, 1);
 
-  free(strings.strings[0]);
-  free(strings.strings);
+  stringfn_release_strings_struct(strings);
 }