+++ tests/test_split_prefix_string.c
@@ -10,9 +10,7 @@ void test_impl()
   assert_string_equal(strings.strings[0], "");
   assert_string_equal(strings.strings[1], "bc");
 
-  free(strings.strings[0]);
-  free(strings.strings[1]);
-  free(strings.strings);
+  stringfn_release_strings_struct(strings);
 }