+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * New stringfn_split_lines_and_trim function.
 * New stringfn_equal function.
 * All trim functions remove all whitespace types characters including: tab, eol and cr.
+* New stringfn_release_strings_struct function.
 
 ### v0.1.0 (2020-11-29)